backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
geotessera>=1.0.0
rasterio>=1.3.0
numpy>=1.24.0
Pillow>=10.0.0
